+++ src/skyrimbsainvalidation.h
@@ -14,7 +14,7 @@ public:
 
 private:
   virtual QString invalidationBSAName() const override;
-  virtual unsigned long bsaVersion() const override;
+  virtual uint32_t bsaVersion() const override;
 };
 
 #endif  // SKYRIMBSAINVALIDATION_H
